Frameworks: Do not allow to access quick settings on secure lockscreen - option (1/2)
We have a lot security guys as user due of using the phones as business
phones...that
said a lot complains go into the direction of google how it is possible
to have
access to quick settings on a secure lockscreen which is absolutely
understandable
due that you are able to control Wifi BT AirplaneMode data etc.

This patch takes care on this problem and adds an option to control this
behaviour.

ATTENTION: Due that this is really an UX derp for an average user this
feature is
enabled by default now. User need to disable it explicit in settings to
get on secure
locksreen the original android behaviour back.

PS2:
- fix derp

Change-Id: Ib437802ae8687dc8e99122a40e285bdd2df3e04c

Signed-off-by: STELIX <ssspinni@gmail.com>